requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
decouple==0.0.7
Deprecated==1.2.13
distro==1.8.0
Django==4.1.3
django-filter==22.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
environ==1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
paramiko==2.12.0
pipreqs==0.4.11
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==5.4.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
texttable==1.6.4
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
yarg==0.1.9
